fix bug: maxcompressLen calc error, UINT32 will overflow
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.6@29780 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/util/utilCompressorLZ4.cpp
@@ -61,12 +61,14 @@ namespace engine
       if ( strategy )
       {
          acceleration = UTIL_COMP_BEST_COMPRESSION - strategy->_level + 1 ;
-         maxExpectedLen = sourceLen * strategy->_minRatio / 100 ;
+         maxExpectedLen = (UINT32)( (UINT64)sourceLen *
+                                    strategy->_minRatio / 100 ) ;
       }
       else
       {
          acceleration = 1 ;
-         maxExpectedLen = sourceLen * UTIL_COMPRESSOR_DFT_MIN_RATIO / 100 ;
+         maxExpectedLen = (UINT32)( (UINT64)sourceLen *
+                                    UTIL_COMPRESSOR_DFT_MIN_RATIO / 100 ) ;
       }
 
       /* The first 4 bytes of the output always contain the original length. */
